--- USB_HID.h
@@ -1,36 +1,0 @@
-#ifndef _USB_HID_
-#define _USB_HID_
-
-#include <avr/power.h>
-#include <avr/wdt.h>
-
-#include "Descriptors.h"
-#include "GamepadEnums.h"
-#include "GamepadDescriptors.h"
-
-static InputMode currentInputMode;
-
-/* Function Prototypes: */
-#ifdef __cplusplus
-extern "C" {
-#endif
-// Setup all necessary hardware, including USB initialization.
-void SetupHardware(InputMode inputMode);
-
-// Process and deliver data from IN and OUT endpoints.
-void HID_Task(void);
-
-// Supported report functions
-void sendReport(void *data, uint8_t size);
-
-// USB device event handlers.
-void EVENT_USB_Device_Connect(void);
-void EVENT_USB_Device_Disconnect(void);
-void EVENT_USB_Device_ConfigurationChanged(void);
-void EVENT_USB_Device_ControlRequest(void);
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif
